Start backend rewrite in Go: config
The old backend still uses Python 2.7, I'm not going to support that on
the new server. But I don't want to migrate to Python 3 and replace
deprecated dependencies, either. Rewriting it in Go feels more feasible.

I'm starting with the configuration. As we're no longer running an
interpreted language, storing the config as a source file is no longer a
viable option, so it will come from environment variables in the future.